+++ CHANGELOG.md
@@ -19,8 +19,10 @@ HumHub Changelog
 - Enh #7317: Space browser: Make the whole space card header and body clickable
 - Enh #7329: Add a new global "Manage Content" Group Permission (see [migration guide](https://github.com/humhub/humhub/blob/develop/MIGRATE-DEV.md#version-117-unreleased) for details)
 - Enh #7325: Add missing IDs in the modal login forms
-- Enh #7333: Improved Yii alias handling and added ENV support  
+- Enh #7333: Improved Yii alias handling and added ENV support
 - Enh #7334: New safe method to rename a database column
+- Enh #7336: Update GitHub workflow versions
+- Enh #7339: Add `DeviceDetectorHelper::isMultiInstanceApp()` method to detect if the app is running in a multi-instance mode
 
 1.17.0-beta.2 (November 12, 2024)
 ---------------------------------